Move DisabledOnSemeru to junit-common
This way, we avoid having two copies of the same code.

(cherry picked from commit 8095bdc7c5c59de4a7eb279f25aa5573d0adc514)
