+++ test-framework/junit-common/src/main/java/io/quarkus/test/junit/common/DisabledOnSemeruCondition.java
@@ -1,4 +1,4 @@
-package io.quarkus.test.junit;
+package io.quarkus.test.junit.common;
 
 import static org.junit.platform.commons.util.AnnotationUtils.findAnnotation;
 
@@ -8,8 +8,6 @@ import java.util.Optional;
 import org.junit.jupiter.api.extension.ConditionEvaluationResult;
 import org.junit.jupiter.api.extension.ExecutionCondition;
 import org.junit.jupiter.api.extension.ExtensionContext;
-
-import io.quarkus.test.common.JdkUtil;
 
 public class DisabledOnSemeruCondition implements ExecutionCondition {
